Renames ncache to cache.
